# src/swagger/swagger.yaml
basePath: /
definitions:
...
  api.ChatPresenceRequest:
    properties:
      chatid:
        description: 'Required: Chat to show typing in'
        type: string
      duration:
        description: 'Optional: Auto-stop after duration (ms)'
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappChatPresenceType'
        description: Text or audio
    type: object
...
  models.QpInviteResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
      url:
        description: invite public link
        type: string
    type: object
...
  models.QpRabbitMQConfig:
    properties:
      broadcasts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle broadcast messages
      calls:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle calls
      connection_string:
        description: RabbitMQ Connection Settings
        type: string
      exchange_name:
        description: RabbitMQ exchange name for routing
        type: string
      extra:
        description: extra info to append on payload
      failure:
        description: Status Tracking
        type: string
      forwardinternal:
        description: Configuration Options
        type: boolean
      groups:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle groups messages
      queue_history:
        description: RabbitMQ history queue name (optional)
        type: string
      readreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit read receipts
      routing_key:
        description: RabbitMQ routing key for exchange routing
        type: string
      success:
        description: last success timestamp
        type: string
      timestamp:
        type: string
      trackid:
        description: identifier of remote system to avoid loop
        type: string
    type: object
  models.QpRabbitMQResponse:
    properties:
      affected:
        description: items affected
        type: integer
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      rabbitmq:
        description: current rabbitmq items
        items:
          $ref: '#/definitions/models.QpRabbitMQConfig'
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  models.QpResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  whatsapp.WhatsappBoolean:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - FalseBooleanType
    - UnSetBooleanType
    - TrueBooleanType
...
  whatsapp.WhatsappChatPresenceType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - WhatsappChatPresenceTypePaused
    - WhatsappChatPresenceTypeText
    - WhatsappChatPresenceTypeAudio
...
info:
  contact:
    name: QuePasa Support
    url: https://github.com/nocodeleaks/quepasa
  description: QuePasa is a Go-based WhatsApp bot platform that exposes HTTP APIs
    for WhatsApp messaging integration
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/nocodeleaks/quepasa/blob/main/LICENSE.md
  termsOfService: https://github.com/nocodeleaks/quepasa
  title: QuePasa WhatsApp API
  version: 4.0.0
paths:
...
  /chat/presence:
    post:
      consumes:
      - application/json
      description: Controls typing indicators and chat presence in WhatsApp conversations
      parameters:
      - description: Chat presence request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Control chat presence
      tags:
      - Chat
...
  /invite:
    get:
      consumes:
      - application/json
      description: Generates an invite link for a specific WhatsApp group
      parameters:
      - description: Chat ID (query parameter)
        in: query
        name: chatid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate group invite link
      tags:
      - Groups
  /invite/{chatid}:
    get:
      consumes:
      - application/json
      description: Generates an invite link for a specific WhatsApp group
      parameters:
      - description: Chat ID (path parameter)
        in: path
        name: chatid
        type: string
      - description: Chat ID (query parameter)
        in: query
        name: chatid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate group invite link
      tags:
      - Groups
...
  /paircode:
    get:
      description: Generates a pairing code for WhatsApp authentication using phone
        number
      parameters:
      - description: Phone number for pairing
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate pairing code
      tags:
      - Connection
  /rabbitmq:
    delete:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpRabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
    get:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpRabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
    post:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpRabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
...
  /scan:
    get:
      description: Generates a QR code for WhatsApp Web authentication scanning
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate QR code
      tags:
      - Connection
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-QUEPASA-TOKEN
    type: apiKey
swagger: "2.0"